requirements.txt
Flask==3.0.2
numpy-stl==3.0.0
numpy==1.26.4
gunicorn==21.2.0
